Move initialize function back into mixins.ts to avoid breaking imports
