+++ lib/contracts/contract.ts
@@ -1,5 +1,3 @@
-import deref = require('json-schema-deref-sync');
-import _ = require('lodash');
 import path = require('path');
 import * as mixins from './mixins';
 import { loadSchemaDefinitionsFromDir } from './utils';
@@ -124,45 +122,6 @@ export interface ContractMap {
 	[slug: string]: ContractDefinition;
 }
 
-interface Mixin<TData = ContractData> {
-	schema: ContractDefinition<TData>;
-	test: (contract: ContractDefinition<TData>) => boolean;
-}
-
-export const initialize = <TData = ContractData>(
-	contract: ContractDefinition<TData>,
-	mixinsToApply: Array<Mixin<TData>>,
-): ContractDefinition<TData> => {
-	const schemasToMerge = [{}, contract];
-
-	mixinsToApply.forEach((el) => {
-		if (el.test(contract)) {
-			schemasToMerge.push(el.schema);
-		}
-	});
-
-	const initializedContract = (_.mergeWith as any)(
-		...schemasToMerge,
-		(objValue: any, srcValue: any) => {
-			if (_.isArray(objValue)) {
-				return _.uniq(objValue.concat(srcValue));
-			}
-			return undefined;
-		},
-	);
-
-	// TODO: This isn't quite right, because the I/O here is contracts not schemas, it works
-	// "by accident" because deref will iterate over any object not just schemas.
-	// Ideally we need to stop using deref completely, in favor of mixin functions, as the
-	// current implementation leads to abominated deep linking.
-	// See https://github.com/product-os/jellyfish-plugin-default/blob/2e15d57ec8b362d899b1957b4ad6fcab5e618b11/lib/cards/mixins/index.js#L12
-	// Dereference all $ref values
-	return deref(initializedContract, {
-		failOnMissing: true,
-		mergeAdditionalProperties: true,
-	}) as ContractDefinition<TData>;
-};
-
 const contracts: ContractDefinition[] = loadSchemaDefinitionsFromDir(
 	path.join(__dirname, '../schemas'),
 	{
@@ -173,7 +132,7 @@ const contracts: ContractDefinition[] = loadSchemaDefinitionsFromDir(
 export const CONTRACTS: ContractMap = contracts.reduce<{
 	[slug: string]: ContractDefinition;
 }>((acc, contract) => {
-	const initializedContract = initialize(contract, [
+	const initializedContract = mixins.initializeContractWithMixins(contract, [
 		{
 			schema: mixins.baseUi,
 			test: (c) => c.type.split('@')[0] === 'type',
